+++ dartsim/src/EntityManagementFeatures.cc
@@ -57,6 +57,8 @@ class ImmobileContactFilter : public dart::collision::BodyNodeCollisionFilter
 
     return BodyNodeCollisionFilter::ignoresCollision(_object1, _object2);
   }
+
+  virtual ~ImmobileContactFilter() = default;
 };
 
 /////////////////////////////////////////////////
@@ -558,7 +560,7 @@ Identity EntityManagementFeatures::ConstructEmptyModel(
         dart::dynamics::Frame::World(),
         _name + "_frame");
 
-  auto [modelID, modelInfo] = this->AddModel({model, modelFrame}, _worldID); // NOLINT
+  auto [modelID, modelInfo] = this->AddModel({model, modelFrame, ""}, _worldID); // NOLINT
 
   return this->GenerateIdentity(modelID, this->models.at(modelID));
 }