+++ dartsim/src/SimulationFeatures_TEST.cc
@@ -41,14 +41,14 @@
 #include <test/PhysicsPluginsList.hh>
 #include <test/Utils.hh>
 
-using TestFeatureList = ignition::physics::FeatureList<
-  ignition::physics::LinkFrameSemantics,
-  ignition::physics::ForwardStep,
-  ignition::physics::GetContactsFromLastStepFeature,
-  ignition::physics::GetEntities,
-  ignition::physics::GetShapeBoundingBox,
-  ignition::physics::sdf::ConstructSdfWorld
->;
+struct TestFeatureList : ignition::physics::FeatureList<
+    ignition::physics::LinkFrameSemantics,
+    ignition::physics::ForwardStep,
+    ignition::physics::GetContactsFromLastStepFeature,
+    ignition::physics::GetEntities,
+    ignition::physics::GetShapeBoundingBox,
+    ignition::physics::sdf::ConstructSdfWorld
+> { };
 
 using TestWorldPtr = ignition::physics::World3dPtr<TestFeatureList>;
 using TestShapePtr = ignition::physics::Shape3dPtr<TestFeatureList>;